+++ src/__init__.py
@@ -1,1 +1,19 @@
-__all__ = []
+__all__ = [
+    'async_client',
+    'client',
+    'metrics',
+    'streaming',
+    'util',
+]
+
+from . import async_client
+from . import client
+from . import metrics
+from . import streaming
+from . import util
+
+try:
+    from . import autoquality
+    __all__.append('autoquality')
+except:
+    pass